transaction

+++ dobriki/activity/models.py
@@ -44,10 +44,10 @@ class ActivityConverter(models.Model):
 
 class ActivityIndicators(models.Model):
     step_count = models.PositiveIntegerField(verbose_name="Количество шагов", default=0)
-    distance = models.DecimalField(max_digits=10, decimal_places=2, verbose_name="Расстояние")
+    distance = models.FloatField(verbose_name="Расстояние")
     time = models.IntegerField(verbose_name="Время")
-    found = models.IntegerField(verbose_name="Найденные бобры")
-    kcal = models.DecimalField(max_digits=10, decimal_places=2, verbose_name="Калории")
+    found = models.FloatField(verbose_name="Найденные бобры")
+    kcal = models.FloatField(verbose_name="Калории")
     activity_type = models.ForeignKey(Activity, on_delete=models.CASCADE)
 
     class Meta:
