+++ dobriki/charity/models.py
@@ -38,7 +38,7 @@ class CharitySubscription(models.Model):
 
 class Wallet(models.Model):
     user = models.OneToOneField(UserModel, related_name='wallet', on_delete=models.CASCADE, verbose_name="Пользователь")
-    balance = models.DecimalField(max_digits=18, decimal_places=2, verbose_name="Баланс", default=0)
+    balance = models.FloatField(verbose_name="Баланс", default=0)
 
     def __str__(self):
         return f"Кошелек {self.user}"
@@ -50,7 +50,7 @@ class Wallet(models.Model):
 
 class Transaction(models.Model):
     user = models.ForeignKey(UserModel, related_name='transactions', on_delete=models.CASCADE, verbose_name="Пользователь")
-    sum = models.DecimalField(max_digits=18, decimal_places=2, verbose_name="Cумма перевода")
+    sum = models.FloatField(verbose_name="Cумма перевода")
     fund = models.OneToOneField(Charity, related_name='transactions', on_delete=models.CASCADE, verbose_name="Фонд")
     created_at = models.DateTimeField(auto_created=True, auto_now=True, verbose_name="Дата проведения транзакции")
